Ajout Validation Champs, Gestion des Exceptions Centralisée, code retour HTTP, Documentation Swagger, pour le microservice patient. Ajout validation pour le microservice  clientui.

+++ patient/src/main/java/com/sante/patient/service/PatientService.java
@@ -1,10 +1,19 @@
 package com.sante.patient.service;
 
+//import static com.sante.patient.constants.Constants.*;
+import static com.sante.patient.constants.Constants.ACTION_CREATE;
+import static com.sante.patient.constants.Constants.ACTION_DELETE;
+import static com.sante.patient.constants.Constants.ACTION_GET;
+import static com.sante.patient.constants.Constants.ACTION_UPDATE;
+
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.sante.patient.exception.PatientNoContentException;
+import com.sante.patient.exception.PatientNotFoundException;
 import com.sante.patient.model.Patient;
 import com.sante.patient.repository.PatientRepository;
 
@@ -16,17 +25,35 @@ public class PatientService implements IPatientService {
 
     @Override
     public Patient createPatient(Patient patient) {
-	return patientRepository.save(patient);
+	Patient patientCreated = patientRepository.save(patient);
+	if (patientCreated == null) {
+	    throw new PatientNoContentException(ACTION_CREATE);
+	} else {
+	    return patientCreated;
+	}
     }
 
     @Override
     public Patient updatePatient(Patient patient) {
-	return patientRepository.save(patient);
+	if (patientRepository.findById(patient.getId()).isEmpty()) {
+	    throw new PatientNotFoundException(ACTION_UPDATE);
+	} else {
+	    Patient patientUpdated = patientRepository.save(patient);
+	    if (patientUpdated == null) {
+		throw new PatientNoContentException(ACTION_UPDATE);
+	    } else {
+		return patientUpdated;
+	    }
+	}
     }
 
     @Override
     public void deletePatient(Long id) {
-	patientRepository.deleteById(id);
+	if (patientRepository.findById(id).isEmpty()) {
+	    throw new PatientNotFoundException(ACTION_DELETE);
+	} else {
+	    patientRepository.deleteById(id);
+	}
     }
 
     @Override
@@ -36,6 +63,11 @@ public class PatientService implements IPatientService {
 
     @Override
     public Patient getPatient(Long id) {
-	return patientRepository.findById(id).get();
+	Optional<Patient> patientOptional = patientRepository.findById(id);
+	if (patientOptional.isEmpty()) {
+	    throw new PatientNotFoundException(ACTION_GET);
+	} else {
+	    return patientOptional.get();
+	}
     }
 }
